feat: make i6 pack encryption optional and take the password with stdin

+++ i6-pack/src/cli.rs
@@ -1,8 +1,31 @@
+use std::io::Write;
+
 use crate::compression;
 use crate::encryption;
 use crate::utils;
 
-pub fn run(action: &str, target: &str, password: &str) -> std::io::Result<()> {
+pub fn run(action: &str, target: &str, encrypt: bool) -> std::io::Result<()> {
+  let password = &if encrypt {
+    print!("Enter password: ");
+    std::io::stdout().flush().expect("Failed to flush stdout");
+    let password1 = rpassword::read_password().expect("Failed to read password");
+
+    if (action == "pack") {
+      print!("Confirm password: ");
+      std::io::stdout().flush().expect("Failed to flush stdout");
+      let password2 = rpassword::read_password().expect("Failed to read password");
+
+      if password1 != password2 {
+        eprintln!("Passwords do not match.");
+        std::process::exit(1);
+      }
+    }
+
+    password1
+  } else {
+    "".to_owned()
+  };
+
   // Validate and sanitize the target path
   let target_path = utils::validate_path(target)
     .or_else(|_| utils::sanitize_output_path(target))
@@ -17,16 +40,26 @@ pub fn run(action: &str, target: &str, password: &str) -> std::io::Result<()> {
   match action {
     "pack" => {
       compression::create_tar_archive(target_path.to_str().unwrap(), tar_file)?;
-      compression::compress_tar_file(tar_file, compressed_file)?;
-      encryption::encrypt_file(compressed_file, encrypted_file, password)?;
+
+      if (encrypt) {
+        compression::compress_tar_file(tar_file, compressed_file)?;
+        encryption::encrypt_file(compressed_file, encrypted_file, password)?;
+      } else {
+        compression::compress_tar_file(tar_file, encrypted_file)?;
+      }
     }
     "unpack" => {
-      encryption::decrypt_file(
-        target_path.to_str().unwrap(),
-        compressed_file,
-        password,
-      )?;
-      compression::decompress_file(compressed_file, tar_file)?;
+      if (encrypt) {
+        encryption::decrypt_file(
+          target_path.to_str().unwrap(),
+          compressed_file,
+          password,
+        )?;
+        compression::decompress_file(compressed_file, tar_file)?;
+      } else {
+        compression::decompress_file(target_path.to_str().unwrap(), tar_file)?;
+      }
+
       compression::extract_tar_archive(
         tar_file,
         &utils::remove_extension(target_path.to_str().unwrap(), ".i6p"),
@@ -40,7 +73,10 @@ pub fn run(action: &str, target: &str, password: &str) -> std::io::Result<()> {
 
   // Clean up temporary files
   std::fs::remove_file(tar_file)?;
-  std::fs::remove_file(compressed_file)?;
+
+  if (encrypt) {
+    std::fs::remove_file(compressed_file)?;
+  }
 
   Ok(())
 }
